+++ usuarios/controladores/modificarUsuario.controlador.js
@@ -1,8 +1,9 @@
 // RF41 Administrador modifica usuario - https://codeandco-wiki.netlify.app/docs/proyectos/tractores/documentacion/requisitos/RF41
 
-const { modificarUsuarioRepositorio: modificarUsuario } = require('../data/repositorios/modificarUsuarioRepositorio.js');
+const { modificarUsuario } = require('../data/repositorios/modificarUsuarioRepositorio.js');
+const { validarYLimpiarUsuario } = require('../../util/servicios/validarUsuario.js');
 const bcrypt = require('bcrypt');
-const validator = require('validator');
+
 
 /**
  * Esta función valida y sanitiza los datos recibidos desde el cuerpo de la petición HTTP.
@@ -17,9 +18,16 @@ const validator = require('validator');
  */
 exports.modificarUsuario = async (peticion, respuesta) => {
     try {
+      if (!peticion.body) {
+        return respuesta.status(400).json({ 
+          mensaje: 'No se recibieron datos para modificar el usuario'
+        });
+      }
+
         const { error, datosSanitizados } = validarYLimpiarUsuario(peticion.body);
 
         if (error) {
+          console.error('Error de validación:', error);
             return respuesta.status(400).json({ mensaje: error });
         }
       
@@ -35,111 +43,23 @@ exports.modificarUsuario = async (peticion, respuesta) => {
         }
         if (idRol) cambios.idRol = idRol;
 
-        await modificarUsuario(idUsuario, cambios);
+        const datos = await modificarUsuario(idUsuario, cambios);
+        if (datos && datos.status) {
+          return respuesta.status(datos.status).json({ mensaje: datos.mensaje });
+        }
     
         return respuesta.status(200).json({ mensaje: 'Usuario modificado exitosamente' });
 
     } catch (error) {
+      if (error.status && error.mensaje) {
+        return respuesta.status(error.status).json({ mensaje: error.mensaje });
+      }
+      if (error.code === 'ER_DUP_ENTRY') {
+        return respuesta.status(400).json({ mensaje: 'El correo ya está en uso' });
+      }
+      
         return respuesta.status(500).json({ mensaje: error });
     }
 }
 
-/**
- * Valida y sanitiza los datos para modificar un usuario
- * @param {object} datos - Datos del usuario
- * @returns {{ error: string|null, datosSanitizados: object|null }}
- */
-function validarYLimpiarUsuario(datos) {
-  const numeroMinimoID = 1;
-  const tamanioMinimoNombre = 1;
-  const tamanioMaximoNombre = 50;
-  const tamanioMinimoCorreo = 5;
-  const tamanioMaximoCorreo = 50;
-  const tamanioMinimoContrasenia = 8;
-  const tamanioMaximoContrasenia = 50;
 
-    const { idUsuario, nombre, correo, contrasenia, idRol } = datos;
-
-    if(!Number.isInteger(idUsuario) || idUsuario < numeroMinimoID) {
-      return {
-        error: 'El identificador de usuario debe ser un número entero mayor o igual a 1.',
-        datosSanitizados: null
-      };
-    }
-
-    // Validar si hay al menos un campo para modificar
-    const existeNombre = typeof nombre  === 'string' && nombre.trim()  !== '';
-    const existeCorreo = typeof correo === 'string' && correo.trim() !== '';
-    const existeContrasenia = typeof contrasenia === 'string' && contrasenia.trim() !== '';
-    const existeRol = idRol != null;
-
-    if(!existeNombre && !existeCorreo && !existeContrasenia && !existeRol) {
-      return {
-        error: 'Debes enviar al menos uno de estos campos para modificar: nombre, correo, contraseña o rol.',
-        datosSanitizados: null
-      };
-    }
-
-    const datosSanitizados = { idUsuario };
-
-    if(existeNombre) {
-      const nombreRecortado = nombre.trim();
-      if(nombreRecortado.length < tamanioMinimoNombre || nombreRecortado.length > tamanioMaximoNombre) {
-        return {
-          error: `El nombre debe tener entre ${tamanioMinimoNombre} y ${tamanioMaximoNombre} caracteres.`,
-          datosSanitizados: null
-        };
-      }
-
-      const regexNombre = /^[A-Za-zÀ-ÖØ-öø-ÿ][A-Za-zÀ-ÖØ-öø-ÿ. ]*$/;
-      if (!regexNombre.test(nombreRecortado)) {
-        return {
-          error: 'El nombre solo puede contener letras (incluidos acentos), espacios y puntos.',
-          datosSanitizados: null
-        };
-      }
-
-      datosSanitizados.nombre = validator.escape(nombreRecortado);
-    }
-
-    if(existeCorreo) {
-      const correoRecortado = correo.trim();
-      if(correoRecortado.length < tamanioMinimoCorreo || correoRecortado.length > tamanioMaximoCorreo) {
-        return {
-          error: `El correo debe tener entre ${tamanioMinimoCorreo} y ${tamanioMaximoCorreo} caracteres.`,
-          datosSanitizados: null
-        };
-      }
-
-      if(!validator.isEmail(correoRecortado)) {
-        return {
-          error: 'El correo electrónico no tiene un formato válido (p. ej. usuario@dominio.com).',
-          datosSanitizados: null
-        };
-      }
-      datosSanitizados.correo = validator.normalizeEmail(correoRecortado);
-    }
-
-    if(existeContrasenia) {
-      const contraseniaRecortada = contrasenia.trim();
-      if(contraseniaRecortada.length < tamanioMinimoContrasenia || contraseniaRecortada.length > tamanioMaximoContrasenia) {
-        return {
-          error: `La contraseña debe tener entre ${tamanioMinimoContrasenia} y ${tamanioMaximoContrasenia} caracteres.`,
-          datosSanitizados: null
-        };
-      }
-      datosSanitizados.contrasenia = contraseniaRecortada;
-    }
-
-    if(existeRol) {
-      if(!Number.isInteger(idRol) || idRol <= numeroMinimoID) {
-        return {
-          error: 'El identificador de rol debe ser un número entero mayor o igual a 1.',
-          datosSanitizados: null
-        };
-      }
-      datosSanitizados.idRol = idRol;
-    }
-
-    return { error: null, datosSanitizados };
-}
